feat: explorer hybrid version
