--- src/app/_components/NavBar/types.ts
@@ -1,9 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface NavItem {
-  id: string;
-  label: string | ReactNode;
-  children?: NavItem[];
-  href?: string;
-  onClick?: () => void;
-}
