+++ src/app/_components/PageTitle.tsx
@@ -5,7 +5,7 @@ import { Heading, HeadingProps } from '../../ui/Heading';
 
 export function PageTitle({ children, ...props }: { children: ReactNode } & HeadingProps) {
   return (
-    <Heading as="h1" fontWeight={'medium'} fontSize="4xl" mb={0} color={'slate.50'} {...props}>
+    <Heading as="h1" fontWeight={'medium'} fontSize="4xl" mb={0} color={'textPrimary'} {...props}>
       {children}
     </Heading>
   );
