+++ src/ui/theme/breakpoints.ts
@@ -1,7 +1,7 @@
 export const BREAKPOINTS = {
-  sm: '30em', // 375px
+  sm: '24em', // 375px
   md: '48em', // 768px
-  lg: '62em', // 1024px
-  xl: '80em', // 1440px
-  '2xl': '96em', // 1920px
+  lg: '64em', // 1024px
+  xl: '90em', // 1440px
+  '2xl': '120em', // 1920px
 };
